Delete user properly

+++ backend/src/main/java/com/google/rolecall/services/UserServices.java
@@ -2,6 +2,7 @@ package com.google.rolecall.services;
 
 import com.google.rolecall.jsonobjects.UserInfo;
 import com.google.rolecall.models.User;
+import com.google.rolecall.repos.CastMemberRepository;
 import com.google.rolecall.repos.UserRepository;
 import com.google.rolecall.restcontrollers.exceptionhandling.RequestExceptions.EntityNotFoundException;
 import com.google.rolecall.restcontrollers.exceptionhandling.RequestExceptions.InvalidParameterException;
@@ -16,6 +17,7 @@ import org.springframework.stereotype.Service;
 public class UserServices {
   
   private final UserRepository userRepo;
+  private final CastMemberRepository castMemberRepo;
 
   public List<User> getAllUsers() {
     List<User> allUsers = new ArrayList<>();
@@ -116,8 +118,13 @@ public class UserServices {
    * @throws EntityNotFoundException The id does not match and existing {@link User}
    *    in the database.
    */
-  public void deleteUser(int id) throws EntityNotFoundException {
-    getUser(id);
+  public void deleteUser(int id) throws EntityNotFoundException, InvalidParameterException {
+    User user = getUser(id);
+
+    if(castMemberRepo.findFirstByUser(user).isPresent()) {
+      throw new InvalidParameterException("User involved in cast or performances cannot be deleted. Change is active.");
+    }
+
     userRepo.deleteById(id);
   }
 
@@ -129,7 +136,8 @@ public class UserServices {
     return Pattern.matches(pattern, email);
   }
 
-  public UserServices(UserRepository userRepo) {
+  public UserServices(UserRepository userRepo, CastMemberRepository castMemberRepo) {
     this.userRepo = userRepo;
+    this.castMemberRepo = castMemberRepo;
   }
 }
